+++ DNBC4tools/DNBC4tools.py
@@ -12,20 +12,20 @@ def main():
     description=textwrap.dedent('''
         DNBC4tools contain the following pipelines and functions:
         --------------------------------
-        pipeline:
-        DNBC4tools run : Run data, count, analysis, report for a complete pipeline.
-        DNBC4tools data : Run quality control and filtering on the raw fastq, 
-                          align the cDNA data to the reference genome and annotate it with gtf file.
-        DNBC4tools count : Determine the inflection point and judge the empty droplets, 
-                           merge multiple beads in the same droplet, calculate the cell * gene expression matrix.
-        DNBC4tools analysis : Run quality control on the cell expression matrix, filter low-quality cells and genes,
-                              perform cell clustering analysis and marker gene screening based on the expression matrix.
-        DNBC4tools report : Summary result files and generate html report.
+        \033[0;32;40mpipeline\033[0m:
+        \033[0;32;40mDNBC4tools run\033[0m      Run data, count, analysis, report for a complete pipeline.
+        \033[0;32;40mDNBC4tools data\033[0m     Run quality control and filtering on the raw fastq, 
+                            align the cDNA data to the reference genome and annotate it with gtf file.
+        \033[0;32;40mDNBC4tools count\033[0m    Determine the inflection point and judge the empty droplets, 
+                            merge multiple beads in the same droplet, calculate the cell * gene expression matrix.
+        \033[0;32;40mDNBC4tools analysis\033[0m Run quality control on the cell expression matrix, filter low-quality cells and genes,
+                            perform cell clustering analysis and marker gene screening based on the expression matrix.
+        \033[0;32;40mDNBC4tools report\033[0m   Summary result files and generate html report.
         
-        function:
-        DNBC4tools multi : Creat shell list for multi samples.
-        DNBC4tools mkref : Create a genome reference directory.
-        DNBC4tools clean : Delete the temp files.'''))
+        \033[0;32;40mfunction\033[0m:
+        \033[0;32;40mDNBC4tools multi\033[0m    Creat shell list for multi samples.
+        \033[0;32;40mDNBC4tools mkref\033[0m    Create a genome reference directory.
+        \033[0;32;40mDNBC4tools clean\033[0m    Delete the temp files.'''))
     parser.add_argument('-v', '--version', action='version', version=_version)
     subparsers = parser.add_subparsers(dest='parser_step')