+++ DNBC4tools/rna/bc_para.py
@@ -8,8 +8,9 @@ parser.add_argument('--cDNAconfig',help='whitelist file in JSON format for cDNA 
 parser.add_argument('--oligofastq1',help='oligoR1 fastq file, Multiple files are separated by commas.',required=True)
 parser.add_argument('--oligofastq2',help='oligoR2 fastq file, Multiple files are separated by commas.',required=True)
 parser.add_argument('--oligoconfig', help='whitelist file in JSON format for oligo fastq.')
-parser.add_argument('--star_index', help='star index dir.')
+parser.add_argument('--genomeDir', help='star index dir.')
 parser.add_argument('--gtf', help='gtf file.')
+parser.add_argument('--adapter', help='adapter file.')
 parser.add_argument('--oligotype', help='oligotype file.')
 parser.add_argument('--thread',type=int, default=4,help='Analysis threads.')
 args = parser.parse_args()
@@ -27,6 +28,7 @@ def cDNA_para():
     cDNA_conf.write('in1=%s/01.data/cDNAin1'%args.outdir+'\n')
     cDNA_conf.write('in2=%s/01.data/cDNAin2'%args.outdir+'\n')
     cDNA_conf.write('config=%s'%args.cDNAconfig+'\n')
+    cDNA_conf.write('adapter=%s'%args.adapter+'\n')
     cDNA_conf.write('cbdis=%s/01.data/cDNA_barcode_counts_raw.txt'%args.outdir+'\n')
     cDNA_conf.write('report=%s/01.data/cDNA_sequencing_report.csv'%args.outdir+'\n')
     cDNA_conf.close()
@@ -55,6 +57,6 @@ def judgeFilexits(*args):
 
 if __name__=='__main__':
     judgeFilexits(args.cDNAfastq1,args.cDNAfastq2,args.oligofastq1,args.oligofastq2,\
-        args.cDNAconfig,args.oligoconfig,args.star_index,args.gtf,args.oligotype)
+        args.cDNAconfig,args.oligoconfig,args.genomeDir,args.gtf,args.oligotype,args.adapter)
     cDNA_para()
     oligo_para()