+++ DNBC4tools/rna/star_ref.py
@@ -1,4 +1,4 @@
-import argparse,os
+import argparse,os,re,collections
 parser = argparse.ArgumentParser(description='Create star index')
 parser.add_argument('--action', metavar='SELECT', default='mkref', choices=['mkref', 'mkgtf', 'stat'], help='Select the action for your program, include mkref,mkgtf,stat, default is mkref') 
 parser.add_argument('--ingtf', metavar='FILE' ,help='Set ingtf in mkref,mkgtf or stat')
@@ -8,8 +8,9 @@ parser.add_argument('--attribute',metavar='DICT',default = ['gene_type=protein_c
 parser.add_argument('--outstat',metavar='FILE', default = 'gtf_type.txt',help='Set the stats outfile in stat, default is "gtf_type.txt" in current dir')
 parser.add_argument('--type',metavar='STR', default = 'gene_type',help='Set the the type for stat, default is gene_type')
 parser.add_argument('--fasta',metavar='FASTA',help='Set the fasta in mkref')
-parser.add_argument('--star_dir',metavar='DIR',default=os.getcwd(), help='Set the star indexdir in mkref, default is current dir')
+parser.add_argument('--genomeDir',metavar='DIR',default=os.getcwd(), help='Set the star indexdir in mkref, default is current dir')
 parser.add_argument('--star',metavar='PROGRAM',help='Set the star program path in mkref')
+parser.add_argument('--limitram',metavar='INT',help='Set the limit genome generate ram in mkref')
 parser.add_argument('--thread',metavar='INT', help='Set the threads in mkref')
 args = parser.parse_args()
 
@@ -45,11 +46,12 @@ def filtergtf(gtf,filtergtf,attribute):
             if i[0].startswith('#'):
                 result.writelines(i)
             else:
-                lst = i[0].split('\t')[-1].replace('"', '').strip().strip(';')
-                lstlist = lst.split('; ')
-                aDict = {}
-                for j in lstlist:
-                    aDict[j.split(' ')[0]] = j.split(' ')[-1]
+                aDict = collections.OrderedDict()
+                pattern = re.compile(r'(\S+?)\s*"(.*?)"')
+                for m in re.finditer(pattern, i[0].split('\t')[-1]):
+                    key = m.group(1)
+                    value = m.group(2)
+                    aDict[key] = value
                 for key1,value1 in aDict.items():
                     for key2,value2 in d.items():
                         if key1 == value2 and key2 == value1:
@@ -68,9 +70,12 @@ def statgtf(gtf,keyword,outfile):
             else:
                 lst = line.split('\t')
                 if lst[2] == 'gene':
-                    lstlist = lst[-1].replace('"', '').strip().strip(';')
-                    llstlist = lstlist.split('; ')
-                    aDict = dict(l.split(' ') for l in llstlist)
+                    aDict = collections.OrderedDict()
+                    pattern = re.compile(r'(\S+?)\s*"(.*?)"')
+                    for m in re.finditer(pattern, lst[-1]):
+                        key = m.group(1)
+                        value = m.group(2)
+                        aDict[key] = value
                     sumDict.append(aDict[keyword])
         result = Counter(sumDict)
         outfile = open(outfile,'w')
@@ -79,15 +84,15 @@ def statgtf(gtf,keyword,outfile):
             outfile.write(f'{k}\t{v}\n')
         outfile.close()
 
-def star_index(fasta,gtf,star_dir,star_program,threads):
-    if not os.path.exists(star_dir):
-        os.system('mkdir -p %s'%star_dir)
-    star_cmd = '%s --runMode genomeGenerate --runThreadN %s --genomeDir %s --genomeFastaFiles %s --sjdbGTFfile %s --sjdbOverhang 99 --limitGenomeGenerateRAM 125000000000'\
-        %(star_program,threads,star_dir,fasta,gtf)
+def star_index(fasta,gtf,genomeDir,star_program,limitram,threads):
+    if not os.path.exists(genomeDir):
+        os.system('mkdir -p %s'%genomeDir)
+    star_cmd = '%s --runMode genomeGenerate --runThreadN %s --genomeDir %s --genomeFastaFiles %s --sjdbGTFfile %s --sjdbOverhang 99 --limitGenomeGenerateRAM %s'\
+        %(star_program,threads,genomeDir,fasta,gtf,limitram)
     print('STAR verison: 2.7.2b')
     print('runMode: genomeGenerate')
     print('runThreadN: %s'%threads)
-    print('genomeDir: %s'%star_dir)
+    print('genomeDir: %s'%genomeDir)
     print('fasta: %s'%fasta)
     print('gtf: %s'%gtf)
     check_call(star_cmd,shell=True)
@@ -98,5 +103,5 @@ if __name__=='__main__':
     if args.action == 'mkgtf':
         filtergtf(args.ingtf,args.outgtf,args.attribute)
     if args.action == 'mkref':
-        star_index(args.fasta,args.ingtf,args.star_dir,args.star,args.thread)
+        star_index(args.fasta,args.ingtf,args.genomeDir,args.star,args.limitram,args.thread)