+++ DNBC4tools/tools/data.py
@@ -10,27 +10,27 @@ class Data:
         self.oligor2 = args.oligofastq2
         self.thread = args.thread
         self.name = args.name
-        self.cDNAconfig = args.cDNAconfig
-        self.oligoconfig = args.oligoconfig
+        self.chemistry = args.chemistry
+        self.darkreaction = args.darkreaction
+        self.customize = args.customize
         self.outdir = os.path.join(args.outdir,args.name)
-        self.starindex = args.starIndexDir
+        self.genomeDir = args.genomeDir
         self.gtf = args.gtf
         self.no_introns = args.no_introns
-        self.mixseq = args.mixseq
 
     def run(self):
-        judgeFilexits(self.cDNAr1,self.cDNAr2,self.oligor1,self.oligor2,self.cDNAconfig,self.oligoconfig,self.starindex,self.gtf)
+        judgeFilexits(self.cDNAr1,self.cDNAr2,self.oligor1,self.oligor2,self.genomeDir,self.gtf)
         str_mkdir('%s/01.data'%self.outdir)
         str_mkdir('%s/log'%self.outdir)
         change_path()
         new_python = python_path()
 
-        scRNA_parse_cmd = ['%s %s/rna/star_anno.py --name %s --cDNAfastq1 %s --cDNAfastq2 %s --oligofastq1 %s --oligofastq2 %s --thread %s --cDNAconfig %s --oligoconfig %s --outdir %s --starIndexDir %s --gtf %s'\
-            %(new_python,_root_dir,self.name,self.cDNAr1,self.cDNAr2,self.oligor1,self.oligor2,self.thread,self.cDNAconfig,self.oligoconfig,self.outdir,self.starindex,self.gtf)]
+        scRNA_parse_cmd = ['%s %s/rna/star_anno.py --name %s --cDNAfastq1 %s --cDNAfastq2 %s --oligofastq1 %s --oligofastq2 %s --thread %s --chemistry %s --darkreaction %s --outdir %s --genomeDir %s --gtf %s'\
+            %(new_python,_root_dir,self.name,self.cDNAr1,self.cDNAr2,self.oligor1,self.oligor2,self.thread,self.chemistry,self.darkreaction,self.outdir,self.genomeDir,self.gtf)]
         if self.no_introns:
             scRNA_parse_cmd += ['--no_introns']
-        if self.mixseq:
-            scRNA_parse_cmd += ['--mixseq']
+        if self.customize:
+            scRNA_parse_cmd += ['--customize %s'%self.customize]
         scRNA_parse_cmd = ' '.join(scRNA_parse_cmd)
         start_print_cmd(scRNA_parse_cmd)
         final_sort_cmd = 'samtools sort -@ %s %s/01.data/final.bam -o %s/01.data/final_sorted.bam'\
@@ -47,17 +47,17 @@ def data(args):
     Data(args).run()
 
 def parse_data(parser):
-    parser.add_argument('--name',metavar='NAME',help='sample name.', type=str)
-    parser.add_argument('--outdir',metavar='DIR',help='output dir, [default is current directory].', default=os.getcwd())
-    parser.add_argument('--cDNAfastq1',metavar='FASTQ',help='cDNAR1 fastq file, Multiple files are separated by commas.', required=True)
-    parser.add_argument('--cDNAfastq2',metavar='FASTQ',help='cDNAR2 fastq file, Multiple files are separated by commas, the files order needs to be consistent with cDNAfastq1.', required=True)
-    parser.add_argument('--cDNAconfig',metavar='JSON',help='whitelist of cell barcode and structure file in JSON format for cDNA fastq.',default='%s/config/DNBelabC4_scRNA_beads_readStructure.json'%_root_dir)
-    parser.add_argument('--oligofastq1',metavar='FASTQ',help='oligoR1 fastq file, Multiple files are separated by commas.',required=True)
-    parser.add_argument('--oligofastq2',metavar='FASTQ',help='oligoR2 fastq file, Multiple files are separated by commas, the files order needs to be consistent with oligofastq1.',required=True)
-    parser.add_argument('--oligoconfig',metavar='JSON',help='whitelist of cell barcode and structure file in JSON format for oligo fastq.',default='%s/config/DNBelabC4_scRNA_oligo_readStructure.json'%_root_dir)
-    parser.add_argument('--thread',metavar='INT',type=int, default=4,help='Analysis threads.')
-    parser.add_argument('--starIndexDir',metavar='DIR',type=str, help='star index dir.')
-    parser.add_argument('--gtf',metavar='GTF',type=str, help='gtf file.')
-    parser.add_argument('--mixseq',action='store_true',help='If cDNA and oligo sequence in one chip, add this parameter.')
+    parser.add_argument('--name', metavar='NAME',help='Sample name.', type=str,required=True)
+    parser.add_argument('--outdir', metavar='PATH',help='Output diretory, [default: current directory].', default=os.getcwd())
+    parser.add_argument('--cDNAfastq1', metavar='FASTQ',help='cDNA R1 fastq file, use commas to separate multiple files.', required=True)
+    parser.add_argument('--cDNAfastq2', metavar='FASTQ',help='cDNA R2 fastq file, use commas to separate multiple files, the files order needs to be consistent with cDNAfastq1.', required=True)
+    parser.add_argument('--oligofastq1', metavar='FASTQ',help='oligo R1 fastq file, use commas to separate multiple files.',required=True)
+    parser.add_argument('--oligofastq2', metavar='FASTQ',help='oligo R2 fastq file, use commas to separate multiple files, the files order needs to be consistent with oligofastq1.',required=True)
+    parser.add_argument('--chemistry',metavar='STR',choices=["scRNAv1HT","scRNAv2HT","auto"],help='Chemistry version. Automatic detection is recommended. If setting, needs to be used with --darkreaction, can be "scRNAv1HT", "scRNAv2HT", [default: auto].',default='auto')
+    parser.add_argument('--darkreaction',metavar='STR',help='Sequencing dark reaction. Automatic detection is recommended. If setting, needs to be used with --chemistry, use comma to separate cDNA and oligo, can be "R1,R1R2", "R1,R1", "unset,unset", [default: auto].', default='auto')
+    parser.add_argument('--customize',metavar='STR',help='Customize files for whitelist and readstructure in JSON format for cDNA and oligo, use comma to separate cDNA and oligo.')
+    parser.add_argument('--thread',type=int, metavar='INT',default=4,help='Number of threads to use, [default: 4].')
+    parser.add_argument('--genomeDir',type=str, metavar='PATH',help='Path of folder containing reference index.',required=True)
+    parser.add_argument('--gtf',type=str, metavar='GTF',help='Path of gtf file.',required=True)
     parser.add_argument('--no_introns', action='store_true',help='Not include intronic reads in count.')
     return parser